- Set Theme to TapPaymentInput - Set Theme to TapSwitch

+++ tapuilibrary/src/main/java/company/tap/tapuilibrary/uikit/views/TapSelectionTabLayout.kt
@@ -366,6 +366,6 @@ class TapSelectionTabLayout(context: Context?, attrs: AttributeSet?) :
 //        const val INVALID_INDICATOR_COLOR = "#a8a8a8"
          val INVALID_INDICATOR_COLOR = Color.parseColor(ThemeManager.getValue("cardPhoneList.underline.unselected.backgroundColor"))
          val UNSELECTED_ALPHA = (ThemeManager.getValue("cardPhoneList.icon.otherSegmentSelected.alpha") as Double).toFloat()
-         val MAX_ITEM_WIDTH = (ThemeManager.getValue("cardPhoneList.maxWidth") as Double).toFloat()
+         val MAX_ITEM_WIDTH = (ThemeManager.getValue("cardPhoneList.maxWidth") as Int).toFloat()
     }
 }
